+++ Program/Spring/demo/demo/src/main/java/com/cedrickcodes/demo/HelloWorldController.java
@@ -1,0 +1,13 @@
+package com.cedrickcodes.demo;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HelloWorldController {
+
+    @GetMapping(path = "/hello")
+    public String HelloWorld(){
+        return "Hello World!";
+    }
+}
