+++ Program/Spring/demo/demo/src/test/java/com/cedrickcodes/demo/DemoApplicationTests.java
@@ -1,0 +1,13 @@
+package com.cedrickcodes.demo;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DemoApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
