+++ Program/TesdaProject/src/FactoryDesign/PokemonData.java
@@ -1,0 +1,10 @@
+package FactoryDesign;
+
+public abstract class PokemonData {
+
+    public abstract int getIdData();
+    public abstract String getName();
+    public abstract String getType();
+
+
+}
