+++ Program/TesdaProject/src/FactoryDesign/WildStats.java
@@ -1,0 +1,13 @@
+package FactoryDesign;
+
+public abstract class WildStats {
+
+    public abstract int getIdStats();
+    public abstract int getHp();
+    public abstract int getAttack();
+    public abstract int getDefense();
+    public abstract int getSpAttack();
+    public abstract int getSpDefense();
+    public abstract int getSpeed();
+
+}
