+++ Program/TesdaProject/src/Main.java
@@ -6,5 +6,6 @@ public class Main {
         //PokemonTemplate template = new Wild();
         //template.buildPokemon();
 
+
     }
 }
